requirements.txt
keras
numpy
tensorflow
flask 
flask_jwt_extended
mysql-connector-python
APScheduler
pillow==10.3.0
chardet